Lots of progress, user authentification

+++ backend/routes/posts.js
@@ -10,6 +10,7 @@ router.post("", checkAuth, (req, res, next) => {
         title: req.body.title,
         content: req.body.content,
         likes: req.body.likes,
+        creator: req.userData.userId
     });
     post.save().then(createdPost => {
         res.status(201).json({
@@ -19,7 +20,7 @@ router.post("", checkAuth, (req, res, next) => {
     });
 });
 
-router.get("",(req, res, next) => {
+router.get("", (req, res, next) => {
    Post.find()
     .then(documents => {
         res.status(200).json({
@@ -46,15 +47,23 @@ router.put("/:id", checkAuth, (req, res, next) => {
         content: req.body.content,
         likes: req.body.likes
     });
-    Post.updateOne({_id: req.params.id}, post).then(result => {
-        res.status(200).json({ message: "update successful" });
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+        if(result.nModified > 0) {
+            res.status(200).json({ message: "Update successful" });
+        } else {
+            res.status(401).json({ message: "Not authorized"});
+        }
+        
     });
 })
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-    Post.deleteOne({ _id: req.params.id}).then(result => {
-        console.log(result);
-        res.status(200).json({ message: "post deleted"});
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId}).then(result => {
+        if(result.n> 0) {
+            res.status(200).json({ message: "Deletion successful" });
+        } else {
+            res.status(401).json({ message: "Not authorized"});
+        }
     });
 });
